requirements.txt
tornado == 4.5.3
SQLAlchemy == 1.2.13
mysqlclient == 1.3.13
requests == 2.20.0
Pillow == 5.3.0
PyJWT == 1.6.4